% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_Photophys.R
\name{calc.ETR_RCII}
\alias{calc.ETR_RCII}
\title{Calculate ETR LabSTAF}
\usage{
calc.ETR_RCII(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
ETR
}
\description{
Calc ETR_RCII
}
