% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_Photophys.R
\name{calc.NPQ}
\alias{calc.NPQ}
\title{Calculate NPQ LabSTAF}
\usage{
calc.NPQ(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
NPQ
}
\description{
LabSTAF NPQ
}
