% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photophys_functions.R
\name{get.Alpha}
\alias{get.Alpha}
\title{Extract Alpha}
\usage{
get.Alpha(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
Alpha
}
\description{
Alpha
}
