% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabSTAF_Notes_functions.R
\name{get.Author}
\alias{get.Author}
\title{Extract author}
\usage{
get.Author(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters- LabSTAF}
}
\value{
Run by
}
\description{
Notes- author
}
