% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabSTAF_Notes_functions.R
\name{get.Blank}
\alias{get.Blank}
\title{Extract Project}
\usage{
get.Blank(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters- LabSTAF}
}
\value{
Blank
}
\description{
Notes- Blank
}
