% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match.sheet.data.R
\name{get.cellcount}
\alias{get.cellcount}
\title{Extract Cells/mL}
\usage{
get.cellcount(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
A numeric vector of cells/mL associated with each file
}
\description{
Extract Cells/mL
}
