% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photophys_functions.R
\name{get.Chl}
\alias{get.Chl}
\title{Extract Chla}
\usage{
get.Chl(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
Chla
}
\description{
Chlorophyll a
}
