% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match.sheet.data.R
\name{get.ChlaExtract}
\alias{get.ChlaExtract}
\title{Extract Chla extract}
\usage{
get.ChlaExtract(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
A numeric vector of Chla extracted from harvest
}
\description{
Extract Chla extract
}
