% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match.sheet.data.R
\name{get.Condition}
\alias{get.Condition}
\title{Extract Condition}
\usage{
get.Condition(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
A character list of light condition
}
\description{
Extract Condition
}
