% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match.sheet.data.R
\name{get.cultureID}
\alias{get.cultureID}
\title{Extract Culture ID}
\usage{
get.cultureID(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
A character list of each files associated culture ID
}
\description{
Extract Culture ID
}
