% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match.sheet.data.R
\name{get.datetime}
\alias{get.datetime}
\title{Extract datetime}
\usage{
get.datetime(frrf)

get.datetime(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
A character list of datetime

A numeric vector of the PAR data associated with the incubation
}
\description{
Extract datetime

Extract PAR
}
