% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photophys_functions.R
\name{get.F}
\alias{get.F}
\title{Extract Fo LabSTAF}
\usage{
get.F(frrf)

get.F(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters- LabSTAF}
}
\value{
Fo

Fo
}
\description{
LabSTAF Fo

F LABSTAF
}
