% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photophys_functions.R
\name{get.File}
\alias{get.File}
\title{Extract File name}
\usage{
get.File(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
File name
}
\description{
File
}
