% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photophys_functions.R
\name{get.fv.fm}
\alias{get.fv.fm}
\title{Extract Fv/Fm}
\usage{
get.fv.fm(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
fv/fm
}
\description{
Fv/Fm
}
