% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photophys_functions.R
\name{get.Fv}
\alias{get.Fv}
\title{Calc and extract Fv}
\usage{
get.Fv(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters- LabSTAF}
}
\value{
Fv
}
\description{
Fv LABSTAF
}
