% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photophys_functions.R
\name{get.fvfm.LS}
\alias{get.fvfm.LS}
\title{Extract Fv/Fm LabSTAF}
\usage{
get.fvfm.LS(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
fv/fm
}
\description{
Fv/Fm
}
