% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photophys_functions.R
\name{get.fvfm}
\alias{get.fvfm}
\title{Extract Fv/Fm Act2Run}
\usage{
get.fvfm(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
fv/fm
}
\description{
Fv/Fm
}
