% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match.sheet.data.R
\name{get.Incubator}
\alias{get.Incubator}
\title{Extract Incubator}
\usage{
get.Incubator(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
A character list of each files associated experimental incubator
}
\description{
Extract Incubator
}
