% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match.sheet.data.R
\name{get.Intensity}
\alias{get.Intensity}
\title{Extract Intensity}
\usage{
get.Intensity(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
A character list of Intensity
}
\description{
Extract Intensity
}
