% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match.sheet.data.R
\name{get.Iron}
\alias{get.Iron}
\title{Extract Iron}
\usage{
get.Iron(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
A character list of Iron level
}
\description{
Extract Iron
}
