% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photophys_functions.R
\name{get.JVPIIm}
\alias{get.JVPIIm}
\title{Extract JVPIIm}
\usage{
get.JVPIIm(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
JVPIIm
}
\description{
JVPIIm
}
