% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match.sheet.data.R
\name{get.LTMHW}
\alias{get.LTMHW}
\title{Extract LTMHW}
\usage{
get.LTMHW(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
A character list of each files associated experimental LTMHW- ASLO 2025, MLB
}
\description{
Extract LTMHW
}
