% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabSTAF_Notes_functions.R
\name{get.Note}
\alias{get.Note}
\title{Extract Note}
\usage{
get.Note(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters- LabSTAF}
}
\value{
Note
}
\description{
Notes- Note
}
