% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photophys_functions.R
\name{get.NPQ.LS}
\alias{get.NPQ.LS}
\title{extract calculated NPQ from LabSTAF}
\usage{
get.NPQ.LS(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
NPQ labSTAF
}
\description{
LabSTAF NPQ
}
