% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabSTAF_Notes_functions.R
\name{get.Project}
\alias{get.Project}
\title{Extract Project}
\usage{
get.Project(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters- LabSTAF}
}
\value{
Project
}
\description{
Notes- Project
}
