% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match.sheet.data.R
\name{get.Replicate}
\alias{get.Replicate}
\title{Extract Replicate}
\usage{
get.Replicate(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
A character list of Replicate
}
\description{
Extract Replicate
}
