% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photophys_functions.R
\name{get.sigma}
\alias{get.sigma}
\title{Extract Sigma}
\usage{
get.sigma(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
sigma
}
\description{
Extract Photophys Data from FRRf files
}
\details{
These functions will extract several base parameters of FRRf data

Sigma
}
