% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photophys_functions.R
\name{get.sigmapii}
\alias{get.sigmapii}
\title{Extract SigmaPII}
\usage{
get.sigmapii(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters- LabSTAF}
}
\value{
sigmaPII
}
\description{
SigmaPII LABSTAF
}
