% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match.sheet.data.R
\name{get.Timepoint}
\alias{get.Timepoint}
\title{Extract Timepoints}
\usage{
get.Timepoint(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
A character list of each files associated experimental timepoint
}
\description{
Extract Timepoints
}
