% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match.sheet.data.R
\name{get.Treatment}
\alias{get.Treatment}
\title{Extract Treatments}
\usage{
get.Treatment(frrf)
}
\arguments{
\item{frrf}{A list of frrf data with components for extracting parameters}
}
\value{
A character list of each files associated experimental treatment
}
\description{
Extract Treatments
}
