% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_FRRf.R
\name{load.frrf}
\alias{load.frrf}
\title{Load FRRF Datafiles}
\usage{
load.frrf(input.dir, file.names = NULL, verbose = T)
}
\arguments{
\item{input.dir}{the input directory}

\item{file.names}{the name of the file in the input directory (default will load all files in that directory)}

\item{verbose}{honestly I have no idea this was all Tom}
}
\description{
http://about.tkelly.org/
}
\author{
Thomas Bryce Kelly
}
