% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Universal_LabSTAF_Loading.R
\name{load.LabSTAF}
\alias{load.LabSTAF}
\title{Load FRRF Datafiles- UNIVERSAL}
\usage{
load.LabSTAF(directory)
}
\arguments{
\item{directory}{Path to the input directory containing .csv files}
}
\description{
Universal LabSTAF Loading
}
\author{
MB
}
