% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_LabSTAF.R
\name{load.LabSTAF_ASLO25}
\alias{load.LabSTAF_ASLO25}
\title{Load FRRF Datafiles - ASLO 2025 MLB}
\usage{
load.LabSTAF_ASLO25(directory)
}
\arguments{
\item{directory}{Path to the input directory containing .csv files}
}
\description{
Load LabSTAF
}
\author{
HoWan Chan
}
