% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match.Files.R
\name{match.files}
\alias{match.files}
\title{Match files}
\usage{
match.files(frrf, Match.sheet)
}
\arguments{
\item{frrf}{FRRf file matching}

\item{Match.sheet}{Corresponding match.sheet for file matching}
}
\value{
fully matched files- appended data from Match.sheet into corresponding FRRf timestamped files
}
\description{
Match files
}
