% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_LabSTAF.R
\name{load.LabSTAF_ASLO25}
\alias{load.LabSTAF_ASLO25}
\title{Load FRRF Datafiles - ASLO 2025 MLB}
\usage{
load.LabSTAF_ASLO25(directory)
}
\arguments{
\item{input.dir}{the input directory}

\item{file.names}{the name of the file in the input directory (default will load all files in that directory)}

\item{verbose}{honestly I have no idea this was all Tom}
}
\description{
Load LabSTAF
}
\author{
HoWan Chan
}
